adds the ability to whitelist the UUID enforcement behavior via an environment variable.
To use it, set the value of "enforce-uuid-whitelist" to a key=value pair, and if that key=value is present in the environment, the UUID enforcement will be skipped (if enabled). This is to allow certain containers, such as the device plugin to continue functioning the way it currently does (by inspecting the NVIDIA_VISIBLE_DEVICES variable).